Help opkg with recipes that have COMPATIBLE_MACHINE
There's a problem in OE's opkg and that's when you use COMPATIBLE_MACHINE on a bb file, BitBake won't allow that recipe to be builded for anything other than what you wrote in COMPATIBLE_MACHINE but the builded ipk file could be installed later via opkg as it has the compatible architecture with some machines.

The only way to bypass this is you use MACHINE_ARCH on any bb file that has COMPATIBLE_MACHINE so opkg won't install it later but this is a quick fix and results in more compiling time and more feed size but if we use common ipk files we could reduce them all.

For example we have 'machine1', 'machine2' and 'machine3' also in the bb file we have:

COMPATIBLE_MACHINE = '^(machine1|machine2|machine3)$'

So BitBake will compile this for the 3 machines but later it could be installed via opkg for other machines too if they have the compatible architecture unless you use MACHINE_ARCH and compile a common packge for each of those 3 machines individually.

If we could have a new field in control file of ipk files called 'Machine' and list those compatible machine opkg could check for:

Machine: machine1, machine2, machine3

In the control file and do not install it for the rest.

We need a new abiltiy in opkg!